vfsd: extract GPT parsing into a host-tested gpt crate
vfsd's GPT header/entry decode and the DPS priority tie-break were entangled
with block-device IPC in a seraph-only binary, so the parse of untrusted
on-disk partition structures had no host coverage. Lift the pure logic into a
new no_std `gpt` crate (parse_header, parse_entry, lookup_partition_by_type_guid)
with 14 value-driven host tests covering signature and entry-size validation,
the unused and inverted-range entry skips, short-slice rejection, and the
priority and tie-break paths. vfsd's src/gpt.rs becomes src/partition.rs: the
block-device I/O that reads sectors and delegates every decode to the crate.

This is the reference pattern for the host-reachable-logic norm in Section D;
sibling extractions for fs/fat, procmgr, memmgr, virtio-core, and ruststd are
tracked separately.

Refs #25
